PaintOnTemplateFeature: Smarter template setup
When a template with our synthetic file name already exists, try hard
to use this template.
When in fact creating a new template image, write the file just to
check writing permissions, but remove it after first loading. Mark
the template and the map dirty. In this way, we won't leave orphan
files when the user chooses to not save the map, or when the app
terminates unexpectedly (crash, Android lifecycle).
Task: GH-1614 (orphan template files)