MapWidget: Take responsibility for template loading
Move responsibility for template file loading on visibility changes
from TemplateTableModel and MapView to MapWidget.

TemplateTableModel is not meant to be more than a GUI adapter for a
map's template state. A MapView is not necessarily connected to a
MapWidget, e.g. early during loading a Map, or during testing. It
still needed hooks to disable eager loading of template files.

MapWidget is actually displaying content. It holds and monitors
an individual MapView which generates visibility change events.
Thus MapWidget is a suitable GUI spot to consolidate the triggering
of loading template files in a lazy manner.