MapEditorController: Fix popup lifetime handling
With this change, popup widget / child widget lifetimes are equally
bound to the map widget in touch mode and in desktop mode.
In mobile mode, the child widget becomes a direct child of the map
widget. In desktop mode, the popup widget / child widget now gets
deleted as soon as the map widget is destroyed.
The inconsistent behaviour was hidden by the fact that the popup
widget methods were used almost exclusively for key button bars,
i.e. in mobile mode.