ScalingIconEngine: New class
For high-dpi screens (not only on mobile phones), we need large icons.
But for most icons, we don't have high-resolution variants.
Instead of scaling the icons (only) in ActionGridBar, we need to do
it in a more central location in order to cover icon usage in regular
toolbars or in the home screen widget.

This commit adds an icon engine which implements scaling up, but
otherwise uses the default Qt engine. The provided plugin claims
the "png" file name extension and relies on the default engine being
selected for the "bmp" file name extension.

Along with the engine, this commit implements changes to the proxy
style which take care of scaling standard icons, and it adds the
now-required ".png" extension to some icon references.