TextObjectAlignmentDockWidget: Replace by QToolBar
In the desktop UI, the toolbar is intentionally added as a floating
QDockWidget: By adding it as a regular toolbar, it is hard do avoid
a major change of the map view geometry.
For the touch UI, the toolbar must be added added at the top of the
screen in order to avoid being covered by the virtual keyboard.
Task: GH-1673 (Refactor custom toolbars)