TemplateAdjustActivity: Revise setup & teardown
There is only a single spot of activation (a button), but teardown
may be initiated by unchecking the button, by closing the dock
widget, by starting another activity, or by deleting the template.
By careful use of signals, slots, and the objects used to control
the lifetime of lambdas listening to signals, we can establish a
safe sequence for teardown.
This change also makes it safe to call the destructor before a call
to init, and it removes the now obsolete dock widget member.
Fixes GH-1690 (crash on macOS when closing the dock widget).