OcdFileExport: Try to avoid absolute template path
When a template is in the map directory, use just the filename instead
of the absolute path. Absolute paths break when moving files between
different systems. The proprietary OCD software can handle filenames
in the map's directory, so we can avoid the user being asked for
looking up the file.