cmake: Relax dependency lookup
Improve developer experience, mostly for MSYS2 (Windows):
Default to activate the embedded build of libpolyclipping
if it is not found in the system and no explicit option
other than 'auto' is given.
Harmonize the lookup of PROJ and GDAL data directories, and
consider the include directories if other sources fail.