Template: Delay setTemplateAreaDirty on setup
When configuring a template for the first time, we don't need to,
and sometimes are unable to properly, determine the template area
immediately after loading: the template area might be adjusted by
the postLoadSetup(), or by setupAndLoad() for center_in_view.

In addition to moving the calls, the Configuring state is now left
unchanged after successful loadTemplateFileImpl(). The responsibility
for the Configuring->Loaded state change moves to setupAndLoad().