Georeferencing: Revise state API
Strictly use an enum value to distinguish local, geospatial, and
invalid geospatial states of georeferencing.
This is complemented by making setState() a private member.

There was no simple test if a georeferencing was geospatial AND valid.
Mapper code had to combine isValid() && !isLocal() to test for this
state. By consistently using the state enum, and adding a state for
invalid geospatial references, code is logically simpler and more
expressive.

Note that the existing state checks were not translated literally,
but adjustments were made where the original seemed to not match
the required context.