TemplateTest: Add GPX on NAD 83 maps
There are issues with NAD 83, in particular the placement of
georeferenced tracks not being consistent with earlier versions
of Mapper v0.8.4 (built with PROJ 4.9.3), and with Mapper v0.9.3.
For now, these tests are marked as expected failures here.

template-track-NA.xmap: GPX once for TemplateTrack, once for GDAL.
template-track-NA-084.xmap: map as created with v0.8.4.
template-track-NA-093-GDAL.xmap: map as created with v0.9.3
  with the track loaded using GDAL.
template-track-NA-093-PROJ.xmap: map as created with v0.9.3
  with GDAL track loading disabled.

template-track-NA-084.xmap is also an example of a map which triggers
the warning about "Georeferencing mismatch" with PROJ >= 6.

Task: GH-1709 (GPX track placement regression)
Co-authored-by: Kai Pastor <dg0yt@darc.de>