FileFormat: Refactor filename extension fixup
Sometimes we need to add a default extension to a new filename. This
change moves existing code from MainWindow to FileFormat. It makes
use of the new function for geospatial vector data export.

The function is modified to work without manipulating the given
list of extensions. It no longer uses the deprecated QRegExp class.