Exporter: Take care of saving template data
Before this change, template data saving was done while saving
template XML configuration. This worked only for the default file
format, and it delayed finishing the writing of the map file until
all modified templates were saved. This behaviour meant increasing
the risk of losing changes to the map, due to crash or forceful
termination (out-of-memory, or Android).

In Exporter::doExport(), it is now easy to catch errors from saving
template data and add messages the to the list of warnings.

Resolves GH-1790 (No error reporting from template saving).
Resolves GH-1791 (Template data saved only for default file format).