file_import_export.h: Add safe assertion facility
Our exporters and importers contain many Q_ASSERT macros which catch
unexpected internal errors in debug builds but have no effect in
release builds. It is a bad idea to let release builds pass these
checkpoints only to run in a crash shortly after. By providing a
custom FILEFORMAT_ASSERT macro, we can turn such errors in exceptions
which are already properly caught. This allows users to save the map
in another format (avoiding the particular error), and to provide more
details when reporting such an issue.

This commit replaces Q_ASSERTS in the central exporter and importer
implementations, but not in the "intrusive" parts of the XML file
format (export/import code embedded in other classes).

By interrupting export/import at former Q_ASSERT statements, this
commit may change behaviour in release builds. However, the way how
it was implemented before was wrong. If the original release build
behaviour is the desired one for a particular spot, this must be
resolved in consistent way.

In addition, there may be a performance penalty for doing checks in
release builds.