OcdFileImport: Ignore some double lines
When "double lines" are not explicitly turned of for a line symbol,
but all related widths are zero, there is no visible result.
There is no significant benefit in preserving all double line
properties by construction of a CombinedSymbol, but there is the
downside of Mapper of having shortcomings in exporting nested
CombinedSymbols (e.g. area with border line). Also, the user interface
is less convenient for CombinedSymbol when the user expects to edit
a simple line.