ProjTransform: Normalize transformations
When using Georeferencing (or ProjTransform) with a "geographic CRS"
in the role of the "projected CRS", always assume unit degrees with
legacy PROJ. This changes the behaviour for transformations done via
legacy PROJ, and so it may affect the positioning of some templates,
but only on the few platforms using legacy version of PROJ (Debian 9,
Debian 10, Ubuntu 18.03).

Transformations via modern PROJ and via GDAL (using any PROJ) remain
unchanged.