KeyValueContainer: Reimplement with std::vector
For object tags, we want to preserve the insertion order in order to
maintain a VCS-friendly file format. It is reasonable to assume that
the number of tags per object is low. (In fact, it is zero for most
objects unless imported from sources such as OSM.) This commit
implements an interface similar to std::map (for insertion and lookup)
for pairs of QString (key, value) on top of std::vector (for order and
locality), and it adjusts all spots which access object tags.
Task: GH-1290 (VCS-friendly file format)