Georeferencing: Add getPreciseDeclination() function
This function returns a convenient value indicating the fully precise
declination information present in the Georeferencing state.  If
passed to 'setDeclination' in the future, then both declination and
grivation will be restored to their present values (assuming
convergence is the same). If convergence has changed, then passing
this value to 'setDeclination' will set grivation to a different value
reflecting the new convergence as well as the full precision of the
former grivation. The declination will be unchanged.

Adding an angle to the returned value provides a declination setting
for a rotated map. Passing this value to 'setDeclination' adjusts
both declination and grivation by the amount of the rotation.