GeoreferencingDialog: Lock grivation with full precision
Add a 'declination_setting' member to the GeoreferencingDialog which
holds the declination and grivation information which was present
at the point when grivation became locked. This enables the grivation
to be revived correctly when unlocking.

Also update test/georeferencing_t's testGrivationAffected, which mimics
the GeoreferencingDialog code. The test no longer expects failures.

The updateGrivation() method is removed because it is no longer used.
The initDeclination() method is removed because it is not used.