Template: Catch some coordinate exceptions
Mapper always saves and loads the alternative template
transformation. But in some cases, files seem to have
transforms with coordinates which were not update properly
when moving the map into the regular  domain. Instead
of not loading these files at all, catch the exception
silently.
There is no evidence that this exception would also occur
for the active transform, so error handling for this data
remains unchanged for now.
Task: GH-1969 ('Coordinates are out-of-bounds' error)