Georeferencing: Intepret 'ortho' projection as spheroidal
When loading a map in XML format, there may be references to
a CRS spec with a PROJ string containing '+proj=ortho'.
Prior to PROJ 7.2 these were handled as spheroidal orthographic
projections. Retain that same interpretation, even if Mapper
is built with PROJ 7.2, by calling 'ensureOrthoIsSpheroidal'
whenever a spec string is loaded from a map.