TextSymbolSettings: Fix Custom tabulator positions output
Custom tabulator positions in most cases were not shown with
a precision of 3 digits. This was caused by using 'g' in the
toString() conversion which considers also the number before the
decimal point. Using 'f' instead fixes the output.
However, since OCAD only uses a precision of 2 digits, the precision
is reduced to 2 digits as well.
The maximum tabulator value is reduced to 500.00mm which is also
the maximum in OCAD.