OcdIcon: Enhance performance of getPaletteColorV9()
Linux perf reports getPaletteColorV9() as one of the hotspots in
file_format_t test. About 17% of CPU time is spent in color
quantization.

The original implementation of getPaletteColorV9() is a straightforward
least-squares match between the examined RGB triplet and the .ocd
palette. The palette contains 125 entries, and the table is searched for
every pixel in the 22x22 px icon. However, on a closer look at the
palette organization, it is clear that there is a correspondence between
the RGB triplet and the palette index.

This patch implements a direct translation from the input RGB triplet to
the palette index and cuts down the run time of `file_format_t
saveAndLoad` from 32 seconds to 27 seconds (-16%) on my system.