OcdFileExport: Fix map notes export to .ocd version 8
If the size of the map notes exceeded 32768 the map notes were
truncated when saving as OCAD 8.
However, instead of truncating to 32767 (the terminating 0 will be
appended) it was truncated to 23767.
When testing the fix with OCAD12 it turned out that it was necessary
to truncate to 32766, otherwise OCAD12 would not open the map.