GeoreferencingDialog: Hide 'Show reference point in' for local CRS
When switching between local and geospatial CRS, the link to
OpenStreetMap was shown even if it was not clickable (for local CRS).
This could also happen for a local CRS when using the Pick on Map and
then pressing Reset.
Besides that there was an orphan 'Show reference point in:' label for
local CRS.
The commit now uses setVisible() for both, the label and the link,
instead of just using setEnabled() on the link.