ScaleMapDialog: Adopt term 'Map reference point'
In the 'Change map scale' dialog, the scaling center can be choosen
to be the 'Georeferencing reference point', even for local CRS.
This is confusing since the reference point is based on map coordinates
(e.g., in mm) and not on geographic coordinates.
In the 'Options' section below the term 'Adjust georeferencing
reference point' also only refers to the coordinates of the map
reference point.
This commit implements Scott Turner's proposal to use 'Map reference
point' instead and Kai Pastor's proposal to use 'Adjust map reference
point' instead.