Adopt term 'Map reference point'
In the 'Rotate map', 'Change map scale' and 'Change scale factor'
dialogs, the rotation/scaling center can be chosen to be
the 'Georeferencing reference point', even for local CRS.
This is confusing since the reference point is based on map coordinates
(e.g., in mm) and not on geographic coordinates.
In the 'Options' section below the term 'Adjust georeferencing
reference point' also only refers to the coordinates of the map
reference point.