OcdFileImport: Refactor parsing of parameter strings
For parsing .ocd parameter strings the core mechanism of parsing the
different strings was present for each of the different strings.
A nested class is added that allows to sequentially parse the string.
For strings that don't contain a 'first' parameter the initial read
position is not moved to the first key/value parameter in order to
allow parsing with 'while (parameters.readNext())' and thus leading to
a similar structure for all parameter string evaluation functions.
The 'first' parameter is not associated with a key. In order to allow
processing the 'first' parameter within the loop, a generic key ('\f')
was added.