ObjectQueryParser: Fix out_of_range access
Avoid an out_of_range access in 'input.at(pos+1)'.
While the Release version of Mapper silently ignores the erroneous
access, the Debug version of Mapper crashes with
"ASSERT: "uint(i) < uint(size())".
The fix in line 861 resolves queries like 'a !' or 'a ~' whereas the
fix in line 883 resolves queries like 'a!' or 'a~'.