ObjectQueryParser: Fix removal of escape chars
Fuction 'fromEscaped()' is the counterpart of 'toEscaped()' to remove
the inserted escape chars '\'.
Whereas it is for 'toEscaped()' necessary to skip over the inserted
escape char by '++i', the additional '++i' is erroneous for
'fromEscaped()' when removing the escape char in case there is a
sequence of escaped characters, e.g. '\"\\'.
After removing the escape char 'i' points to the escaped char and the
loop's '++i' will correctly move 'i' to the next escaped or not
escaped char.