XMLFileImporter: Make scale factor check contingent on PROJ version
The 'validateGeoreferencing' method checks for inconsistent
scale factors that arise from a change to PROJ that appeared
in version 7.2.

This change suppresses that check if this Mapper was built with
an earlier version of PROJ. The PROJ version is identified by
CMake and passed to the compiler in the PROJ_PULL_2361 macro.