XMLFileImporter: Check for changed meaning of 'ortho' projection
With PROJ 7.2, 'ortho' supports ellipsoidal variation of the
projection, changing the meaning of some spec strings, and
upsetting the map's georeferencing. Generally, this makes the
georeferencing's scale factor properties inconsistent with the
new interpretation of the projection.

In this change, the 'validate_georeferencing' method is extended to
detect this situation and add "+f=0" to the spec string. "+f=0" makes
explicit that the projection is spheroidal, restoring the effective
projection to what it was when the map was originally georeferenced
with an earlier version of PROJ.