PointSymbolEditorWidget: Don't duplicate CurveStart property
When adding a path point by clicking on the '+' button either a default
point is appended to the path or an existing path point is duplicated.
However, when duplicating an existing point, also the CurveStart
property was duplicated which lead to two successive path points with
the CurveStart property being set although the next two path points
following a CurveStart path point define the course of the curve and
thus can't be a CurveStart path point itself.

Closes GH-2090 (Symbol Editor: Cannot remove coordinate pairs).