MapEditorController: Select pan tool on touch UI startup
The edit mode is selected by default on editor startup. However, in the
touch UI, object editing is rarely the first action to be done after
opening a map. The user typically pans the map view to the area of
interest first and then starts editing.

This code change makes the touch UI start in the view pan mode. The
desktop action remains to be editing as the desktop users use mouse
middle button for editing.

Discussed in GH-982.