OcdFileImport: Import of image objects
Image objects (as well as layout objects) define their color directly
in the object definition.
With this commit the import of layout, image and graphic objects is
now done within a single function.
A single warning is issued that these kind of objects are imported by
means of auxiliary symbols and can't be exported in their original
format.
Closes #1231