OcdFileImport: Set cap and join styles in separate function
Setting of cap and join styles was done in the importLineSymbolBase()
function and was replicated by this PR for importing special objects.
In addition setupLineSymbolFraming() was setting the cap and join
styles for framing lines (although only checking for a subset of cap
and join styles).
This commit moves the setting to a separate function.