OcdFileImport: Import of image layout objects
Image layout objects are stored in parameter strings very similar to
background images (i.e., templates), whereas the image file itself is
stored outside the .ocd file.
This commit imports image layout objects as templates (and informs the
user about it) after the original templates have been imported and
thus adds them to the end of the list of templates.