ColorListWidget: Show usage of color before deletion
When the user decided to delete a color that was used by at least one
symbol, a warning was issued: "The map contains symbols with this color.
Deleting it will remove the color from these objects!".
This warning was however anbiguous: it was even issued when the
affected symbol(s) were not used by any object. In addition, the
mixture of "symbols" and "these objects" sounded awkward.
With this commit the number of affected symbols is counted and put out
in the warning message.
In addition the number of objects actually using these affected symbols
is counted as well. If there is at least one object then the warning is
extended by stating the number of affected objects. In any case the
message text is changed to "...remove the color from these symbols".