ColorListWidget: Show affected symbols before deletion
This commit allows to select the symbols that will be affected by
removing a given color by extending the warning message box by a
new button 'Select affected symbols and abort'.
Pressing on this button then selects all of the affected symbols
in the symbol window.
The SymbolRenderWidget class was extended by selectMultipleSymbols()
to select a number of symbols, this was only possible for a single
symbol before. selectMultipleSymbols() itself is called by the new
method selectMultipleSymbols() in the SymbolWidget class.
Since the ColorListWidget object can neither directly access the
SymbolWidget object nor the parent MapEditorController object which
owns and does provide access to SymbolWidget, MapEditorController now
passes a pointer to himself to ColorListWidget's constructor.

Closes #2122