OgrFileImport: Improve DXF import of text objects
Text object properties like the text itself, alignment and rotation were
imported by putting them into a string and transporting the string via
the description property of the related text symbol.
This approach suffers from multiple deficiencies: the auxiliary string
was not removed from the description field, the rotation angle was
rounded (e.g., -17.2 became 20) and the way of constructing and parsing
the string was unnecessary complex.
Use the auxiliary symbol properties instead to transport the object
properties.