OgrFileImport: Improve DXF import of text objects
Text object properties like the text itsef, alignment and rotation were
previously imported by putting them into a string and transporting the
string via the description property of the related text symbol.
Besides being an awkward design the approach suffers from multiple
deficiencies: the auxiliary string was not removed from the description
field, the rotation angle was rounded (e.g., -17.2 became 20) and the
way of constructing and parsing the string was unnecessary complex.

This commit uses the auxiliary symbol properties to transport the
object properties.