Auxiliary symbol properties (GH-2136)
When importing files, Mapper may encounter symbol properties that are
not part of its native symbol properties but need to be applied when
importing objects.
For example, horizontal and vertical text alignment are symbol
properties in OCAD but object properties in Mapper.
Extend the Symbol base class to support auxiliary symbol properties
that are neither saved to a map file nor loaded from it.
Use auxiliary symbol properties for the GDAL importer and for importing
.ocd files.