Symbol: Auxiliary symbol properties
Mapper sometimes needs to handle symbol properties that are not part of
Mapper's own set of symbol properties and that need to be applied when
importing objects.
Example: horizontal and vertical text alignment is a symbol property in
OCAD but an object property in Mapper.
Mapper uses auxiliary data structures to temporarily store those symbol
properties for applying them later on object import.
This change adds a QVariantHash element to the Symbol base class that
allows a flexible and type safe storage of auxiliary symbol properties.
These auxiliary symbol properties are neither saved to a map file nor
loaded from it.