Symbol: Auxiliary symbol properties
Mapper sometimes needs to handle symbol properties that are not part of
Mapper's own set of symbol properties and that need to be applied when
importing objects.
Example: horizontal and vertical text alignment is a symbol property in
OCAD but an object property in Mapper.
Extend the Symbol base class for storing auxiliary symbol properties
that are neither saved to a map file nor loaded from it.