OcdFileImport: Handle emtpy area holes
There are .ocd files that erroneously contain an empty area hole, i.e.,
there are two subsequent points with the Ocd::OcdPoint32::FlagHole
property.
If such a malformed .ocd file is imported Mapper will crash if the
area object containing the empty hole is selected.
To fix this, the first of these .ocd points needs to be ignored.
Since setPointFlags() applies the HolePoint property to the last point
instead of the first point of the next part, the consequence is that
instead of removing the first point of these two .ocd points, the
second point of the two Mapper points with the HolePoint property needs
to be removed.
Using a loop ensures that multiple subsequent empty area holes are
removed.

Co-authored-by: lpechacek <lpechacek@gmx.com>