OcdFileImport/Export: Maintain v12 object timestamps
Starting from v12 format the object features fields for creation and
modification times. Some users use the modification time to identify
changed objects. Mapper can support this workflow by maintaining the
creation and modification times on the map objects. This patch does not
implement object timestamp updates in Mapper.

The conversion of the floating point number to date and time is
described at
https://docwiki.embarcadero.com/Libraries/Alexandria/en/System.TDateTime