Consider dash symbols in combined symbols
When drawing line objects for symbols with dash symbols, no dash
points were set automatically if the symbol was a combined symbol.
When editing line objects for symbols with dash symbols, additional
points were not added as dash points if the symbol was a combined
symbol.
These deviations were known and marked as TODO.
This commit treats line symbols with dash symbols and combined symbols
containing line symbols with dash symbols equally.