JSON based Mapper configuration
In addition to the configuration of Mapper through the 'Settings' GUI
where a dedicated control element is needed for each setting, this
commit implements a lightweight mechanism by using a JSON configuration
file that allows to easily configure a variety of settings.
The 'Settings' GUI is extended by the possibility to select and
remember the JSON configuration to be used as default configuration.
An additional dialog allows to load and save JSON configuration and
provides a simple, i.e., not context-sensitive, text editor.
The validity of the JSON format is checked and deviations to Mapper's
default JSON configuration can be shown and resolved.