Georeferencing: Use new geographic CRS for pivot, tracks and more
Replace Georeferencing::ballpark_geographic_crs_spec with
Georeferencing::gns_crs_spec in all occurrences that relate to Mapper
internals. The former refers to the long-term evolution of WGS84,
while the latter to an up-to-date realization G1762.

Georeferencing::ballpark_geographic_crs_spec continues to be used
for the map file format, for interoperability with older versions
of Mapper.