TemplateTest: Update expectations for WGS84 improvements
With Mapper now using the WGS84 (G2139) realization for templates and
as a pivot for coordinate transformations, template alignment has
changed, causing differences in the test results. The test cases
    TemplateTrack NAD83
    OgrTemplate NAD83
in ogrTemplateTest no longer fail, because the tracks are now aligned
well enough with the map and its datum to pass the test. In addition,
TemplateTrack and OgrTemplate now align tracks the same, causing those
same test cases to pass in the templateTypesConsistentTest.

On the other hand, the test cases
    TemplateTrack from v0.8.4
    OGRTemplate from v0.9.3
in ogrTemplateTest now fail, because behavior of earlier versions of
Mapper did not handle track coordinates as accurately. The position
is changed by 0.3mm on the map, or 1.3m on the ground.