Georeferencing: Use EPSG:9057 as the GNSS geographic CRS
This changes Georeferencing::gnss_crs_spec from EPSG:9755 to
EPSG:9057. PROJ handles EPSG:9057 with with an epoch several years
earlier than EPSG:9755, which makes coordinate transformations to
NAD83 less accurate, by 20 cm.

With EPSG:9755, Mapper was failing tests in Azure pipelines.
Changing to the earlier EPSG:9057, which indicates the
WGS 84 (G1762) realization, may work with the older version
of PROJ that's available in the superbuild.