Georeferencing: Make geographic CRS a member of ProjTransform
Each Georeferencing object is given its own geographic_crs value,
so that Mapper will be able to support older maps without affecting the
alignment of their templates, while simultaneously providing more
accurate positioning of georeferenced templates in new maps.

For this purpose, ProjTransform objects are enhanced to have an
explicit geographic CRS in addition to the projected CRS.

This modification to ProjTransform is unsuited to another scenario to
which ProjTransform used to be applied, namely when the ProjTransform
would do no more than indicate a CRS. The mechanism supporting
'forward' and 'inverse' methods of such objects was not constructed. A
new struct, 'ProjCRS' is provided for this specialized need. Splitting
off ProjCRS from ProjTransform provides clarity to see which of these
utility objects define a transformation from a geographic CRS, and
which define just a CRS.