Georeferencing: Add getGeographicCRSSpec() method
The static value Georeferencing::gnss_crs_spec is made private to the
Georeferencing class, so that non-core uses of the map's geographic
WGS84 CRS will be sure to depend on the map to obtain the CRS, rather
than specify a particular variant of WGS84. Their references to
gnss_crs_spec are replaced with calls to getGeographicCRSSpec(), which
in the future will depend on the map. The Georeferencing's
toGeographicCoords, toProjectedCoords, and toMapCoords methods will be
consistent with the CRS returned by getGeographicCRSSpec().