Georeferencing: Enable datum compatibility mode
The attribute 'is_realization' is added to the Georeferencing class
and to the geographic_crs_spec of the xml file format. A 'true' value
indicates the new, more accurate geographic CRS, while 'false'
provides the same geographic CRS as earlier versions of Mapper.  The
geographic CRS is used as pivot when transforming from one version of
map coordinates to another. It is also used for GNSS tracks and for
display of geographic coordinates.

The Georeferencing::setDatumBallpark() method provides for the
compatibility mode to be turned on or off.

On loading a map, the user is warned if they haven't decided whether
compatibility is wanted. When the user opens the georeferencing dialog,
a message box asks the user whether they want extra accuracy, or
alignment that is compatible with older Mappers.

Example and test maps are updated to indicate the 'is_realization'
attribute. The template_t test had two test files as examples of what
older versions of Mapper would save. This commit provides the
compatibility which makes the corresponding two tests succeed.