OcdFileExport: Ensure null-terminated text objects
When exporting text objects the terminating null character was missing
if the text length was exactly a multiple of 32 characters.

The .ocd format recommends that every string is null-terminated
although OCAD is tolerant against missing null-termination.
As of 2024 Mapper and Condes however are not tolerant against missing
null-termination.

Another PR will harden the .ocd import by coping with missing
null-termination.
Fixes #2206