MapEditorController: Disable overprinting when spot colors are gone
It is possible to enable overprinting simulation and then delete all
spot colors. The result is no map display and disabled overprinting
simulation option in the menu. The user is locked out of the map content
until addition of a new spot color to enable the menu item.  This patch
makes Mapper react on the absence of spot colors by disabling the
overprinting effect.

Side remark: MapEditorController::newCheckAction connects the action's
triggered() signal to the target slot.  Qt documentation clearly states
that the signal is not emitted upon setChecked() call.  If
newCheckAction linked to the toggled() or changed() signals, the call
added by this patch might be superfluous.