Licensing: Consider derivative distributions
The $ID_LIKE variable in /etc/os-release file may list space separated
names that the current distribution is similar to. For example in some
random distribution that derives from Ubuntu the values could be:

    ID="random-distro"
    ID_LIKE="ubuntu debian"

When trying to guess the current system, use all those values in order
and use the first one that is known.