DrawPathTool: Highlight object that is going to be followed
In complex map situations it is unclear which object will be followed
when the user uses the Shift modifier key to follow lines or area object
boundaries. This change makes the object selection more obvious.

As part of the job we are hardening and making more use of
finishFollowing() that removes the highlight renderables for us.

Co-authored-by: Matthias Kuehlewein <github@kuehlewein.de>