PathCoordVector: Align update with recalculateParts
If input data has invalid hole points, do not prepend them to the
following path part, but create an empty path part instead.

This prevents crashes in locations where the part end is expected
to never be before path start. The empty path part might be another
source of problems, but it is an artifact which is already produced
by recalculateParts() (incl. the PathObject constructor).