MapColor, XMLFile: Introduce color id
The color id value will help us reference the map color in file format
context. Currently colors are addressed by its rendering priority. The
goal is to address colors by its stable id in a future file format.

For now, we do not compare the color id in MapColor::equals() as there
is no use for it. The id is not part of the color identity, much like
its priority says nothing about the color itself.