OcdFileExport: Export non-integer CMYK components
CMYK color component values can include fractional parts. The v8 format
expresses the component values as a number in the range 0-200 that maps
to percentages therefore component values are recorded with 0.5 %
precision. The higher format versions use strings for the color
definition, and the numbers can include one decimal place. This patch
makes Mapper export the fractional part as needed.