MapEditorController: Disable menu item if no symbols are selected
The 'Select all objects with selected symbols' menu item below the
'Edit' menu was enabled even if no symbols were selected.
The same menu item in the popup symbol menu of the symbol pane however
was disabled if no symbols were selected.
Move the enabling or disabling of the menu item to a function which
is called by a change of the selected symbols and disable the menu item
if no symbols are selected.