OcdFileExport: Fix OCD v8 object size calculation
Objects with a large number of coordinates overflow the 16-bit integer
allocated for the object size. As a result, Mapper produced invalid
entries for these objects.

The actual number of object coordinates is limited by the object's
num_items data entry which is indeed a signed short integer. We are now
checking for this condition and refusing to export these objects as
they are invalid in the context of this file format.