MapEditorController: Use standard buttons for map notes dialog
The map notes dialog had a non-standard button layout (with the
Cancel button on the left side and the OK button on the right side).
In addition both buttons were non-standard buttons (the OK button
contained a green right arrow as icon).
Change the map notes dialog layout to use standard buttons as for all
other dialogs in Mapper.