Refactor distribute points tool
The core functionality to distribute points evenly along a path was
partly contained in the MapEditorController class (which was marked as
a TODO).
Move the functionality to the (renamed) DistributePointsDialog class.
Add three dots to the submenu item's name to indicate that selecting it
will open another window.
Remove the 'Settings' since changing them in the dialog was not made
persistent.