Map: Fix deleting object selection
Deleting a selection of objects lead to a corrupted undo step
which could lead to a failure when undoing the undo step before.
The reason was that the selected objects were added in a loop to the
undo step including their object index. After adding an object it was
removed which caused the indexes of the remaining (and sometimes to
be removed as well) objects to be renumbered.
Use the AddObjectsUndoStep::removeContainedObjects() function to
remove the objects once they all have been added.

Closes GH-2320 (Deleting objects corrupts undo step).