OgrFileImport: Import rotation angle of point objects (DXF)
The DXF driver of GDAL/OGR does not (yet) import the rotation angle
of point objects (code 50).
However, the driver stores all codes which are not handled by the
driver itself in a generic field 'RawCodeValues' if the driver is
configured with DXF_INCLUDE_RAW_CODE_VALUES being set to TRUE.
Parse the 'RawCodeValues' field for the occurrence of a code 50
property and set the rotation angle accordingly.

Closes GH-2133 (Point object angle/rotation is lost during DXF import).