AreaSymbol: Adopt 'rotatable' property
This property already exist for Symbol, but wasn't used in AreaSymbol:
area symbol fill patterns carry individual rotatability properties.
The new symbol property does not interfere with the pattern property:
Its role is to provide a hint to map rotation operations, as
implemented in the OC*D software. (However, the effect is not yet
implemented here.)
Existing maps are not affected because the existing rendering and
editing behavior doesn't need to care for this property. Old versions
of Mapper would drop the "rotatable" property, but this doesn't have an
immediate effect until exporting to OCD or (in the future) rotating the
map. Given this low impact, this PR does not add a omap format version.